feat: impl send_formatted
